added hpc evaluation
modified the original DSD_eval to run on the HPC (high performance cluster). this requires another .sh file which calles the .m for each of the files 1:100. each case in then launched on a separate node of the HPC.
to gather the results in a single .mat file you can use DSD100_gather_hpc.m